backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
openrouter-api
openai